+++ app/shared/src/androidMain/kotlin/ui/subject/episode/details/EpisodeDetails.android.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.unit.dp
 import me.him188.ani.app.data.models.subject.SubjectInfo
 import me.him188.ani.app.data.source.media.EpisodeCacheStatus
-import me.him188.ani.app.session.AuthState
+import me.him188.ani.app.data.source.session.AuthState
 import me.him188.ani.app.ui.foundation.ProvideCompositionLocalsForPreview
 import me.him188.ani.app.ui.subject.cache.TestMediaList
 import me.him188.ani.app.ui.subject.collection.EditableSubjectCollectionTypeState
