+++ app/shared/src/commonMain/kotlin/data/repository/TokenRepository.kt
@@ -7,8 +7,8 @@ import androidx.datastore.preferences.core.longPreferencesKey
 import androidx.datastore.preferences.core.stringPreferencesKey
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
-import me.him188.ani.app.session.Session
-import me.him188.ani.app.session.SessionManager
+import me.him188.ani.app.data.source.session.SessionManager
+import kotlin.time.Duration.Companion.hours
 
 /**
  * Do not access directly. Use [SessionManager] instead.
@@ -25,6 +25,17 @@ interface TokenRepository : Repository {
 
     suspend fun clear()
 }
+
+/**
+ * Bangumi 账号信息.
+ */
+data class Session(
+    val accessToken: String,
+    val expiresAtMillis: Long,
+)
+
+fun Session.isValid() = !isExpired()
+fun Session.isExpired() = expiresAtMillis <= System.currentTimeMillis() + 1.hours.inWholeMilliseconds
 
 internal class TokenRepositoryImpl(
     store: DataStore<Preferences>,
