+++ app/shared/src/commonMain/kotlin/data/source/DanmakuManager.kt
@@ -14,9 +14,9 @@ import kotlinx.coroutines.flow.retry
 import kotlinx.coroutines.isActive
 import kotlinx.coroutines.withTimeout
 import me.him188.ani.app.data.repository.SettingsRepository
+import me.him188.ani.app.data.source.session.SessionManager
+import me.him188.ani.app.data.source.session.verifiedAccessToken
 import me.him188.ani.app.platform.getAniUserAgent
-import me.him188.ani.app.session.SessionManager
-import me.him188.ani.app.session.verifiedAccessToken
 import me.him188.ani.app.ui.foundation.BackgroundScope
 import me.him188.ani.app.ui.foundation.HasBackgroundScope
 import me.him188.ani.danmaku.ani.client.AniDanmakuProvider
