+++ app/shared/src/commonMain/kotlin/ui/foundation/AbstractViewModel.kt
@@ -26,7 +26,7 @@ import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.cancel
-import me.him188.ani.app.session.AuthorizationCancelledException
+import me.him188.ani.app.data.source.session.AuthorizationCancelledException
 import me.him188.ani.utils.logging.debug
 import me.him188.ani.utils.logging.error
 import me.him188.ani.utils.logging.logger
