+++ app/shared/src/commonMain/kotlin/ui/profile/AccountViewModel.kt
@@ -5,9 +5,9 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import me.him188.ani.app.data.repository.ProfileRepository
-import me.him188.ani.app.session.AuthState
-import me.him188.ani.app.session.SessionManager
-import me.him188.ani.app.session.userInfo
+import me.him188.ani.app.data.source.session.AuthState
+import me.him188.ani.app.data.source.session.SessionManager
+import me.him188.ani.app.data.source.session.userInfo
 import me.him188.ani.app.ui.foundation.AbstractViewModel
 import me.him188.ani.app.ui.foundation.launchInBackground
 import org.koin.core.component.KoinComponent
